Add files via upload
Description:
This is a simple GUI program written in Python using the Tkinter library for creating a to-do list. The program allows the user to add tasks to a list, remove tasks, and save the list to a JSON file. The GUI is made up of a text entry box, a "Add Task" button, a "Remove Task" button, and a Listbox widget for displaying the current tasks.

The program demonstrates basic concepts of Python and Tkinter, such as reading and writing to files, handling user inputs, and updating the GUI dynamically. It can be used as a starting point for building more complex to-do list applications or as a learning resource for beginners to Python GUI programming.

Features:

Add new tasks to the list
Remove tasks from the list
Save the task list to a JSON file
Load the task list from a JSON file
Simple and intuitive GUI interface
Usage:

Clone or download the repository to your local machine.
Install Python and Tkinter library (if not already installed).
Run the program using the command: python todo_list.py
Add tasks to the list by typing in the text entry box and clicking the "Add Task" button.
Remove tasks by selecting them in the Listbox and clicking the "Remove Task" button.
Save the task list to a JSON file by clicking the "Save" button.
Load a saved task list by clicking the "Load" button and selecting the JSON file.
Contributing:
Contributions are welcome and encouraged! Feel free to open an issue or pull request if you find a bug or want to suggest a new feature. Please follow the standard GitHub workflow and code of conduct.